oss_manager_pyqt/pyproject.toml
[project]
name = "oss-manager-pyqt"
version = "0.1.4"
description = "Standalone PyQt OSS history manager for Git pointer-ref asset repositories"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyQt6>=6.5",
  "PyYAML>=6.0",
  "boto3>=1.34",
]

[project.optional-dependencies]
aliyun = ["oss2>=2.18"]
s3 = ["boto3>=1.34"]

[project.scripts]
oss-manager = "oss_manager.__main__:main"
program = "oss_manager.__main__:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
